Stops modules from loosing their item, enhances a check.
The item rig_modules/device uses gets the ABSTRACT flag added, which
should indicate every object that get attackby with the item to not take
it. The check if we even have a rig part is for suits without all parts,
the internal affairs suit, for example, to stop it from throwing the
message.