Fixing the rig-system (#21)
* Fixes the mounted gun and magboots

Switches from process_fire to after_attack to let the gun handle
everything.
Dublicates the magboots sprites to allow on/off sprites.
Forces magboots to be off while the suit is unsealed.

* Depowers magboots with the suit

Just a small check when the suit goes offline to disable the magboots
like the unsealing. Still needed: Boots stay offline.

* Fixes a minor typo

The race is referenced in the code as Tajaran, not Tajara, which caused
the sprites to not load correctly.

* First batch of racial sprites with headlamps

Headlamps work as far as i can tell, no kind of disabling while
unsealing/loosing power yes, sprites for Tajaran, Human, Unathi, special
Skrell helmets

* Magboots/Headlamp same deactivation, minor runtime fix.

* Stops stripping of suit parts, adds the obj icons

* Adds Rig removal surgery/more disabling of boots/helmet

The rig surgery code is mostly copy-paste from Bay12, Light/Magboots get
disabled on a suit reset, Light can't be toggled while unsealed anymore.

* Disables the disabling on no-power

That way, the Light/Boots work more consistent, the code should be
consistent enough to play with them now.

* Stops modules from loosing their item, enhances a check.

The item rig_modules/device uses gets the ABSTRACT flag added, which
should indicate every object that get attackby with the item to not take
it. The check if we even have a rig part is for suits without all parts,
the internal affairs suit, for example, to stop it from throwing the
message.

* Small changes to the rig surgery

* Fixes a small mistake.

Normally, you would say, shoes are the same as boots, just not when it's
about the name of a variable.