Fixes up rig removal surgery
Dropped sends in the wearer as user, and since stripping can easily mean
the wearer/user is incapacitated, the removal of the rig didn't take the
parts. Big diff in rig_removal.dm as there was a wrong indentation.
Added the correct returns for the surgery to finish and a name to get
shown as step.